Factor out database reader in webserver.
Taking advantage of async_trait in tide::utils
(which comes from async-trait crate).
