+++ bughouse_webserver/Cargo.toml
@@ -11,6 +11,7 @@ clap = { version = "4.0.29", features = ["cargo", "derive"] }
 femme = "2.2.1"
 humantime = "2.1.0"
 log = "0.4.17"
+plotly = "0.8.3"
 skillratings = "0.22.0"
 sqlx = { version = "0.6.2", features = ["postgres", "sqlite", "runtime-async-std-rustls", "time"] }
 tide = "0.16.0"
@@ -18,4 +19,3 @@ tide-jsx = { version = "0.4.0" }
 time = { version = "0.3.17", features = ["formatting"] }
 
 bughouse_chess = { path = ".." }
-plotly = "0.8.3"
