+++ bughouse_webserver/src/lib.rs
@@ -12,6 +12,7 @@ use time::OffsetDateTime;
 
 use bughouse_chess::persistence::GameResultRow;
 
+pub mod database;
 pub mod history;
 
 type Rating = WengLinRating;
